chore: Move the Serde impl for `IpAddr` to a `utils` module
