+++ tests/schemas/utils/mod.rs
@@ -1,0 +1,3 @@
+//! Utilities needed for testing generated code.
+
+pub mod ip_addr_serde;
